Blind reacts on rain, but still several issues
Now the blind is moving up on a rain event. But if the current state is
JT_ON then the blind doesn't move. If it is in JT_OFF then it does.
Also it reacts on "no rain" closing the blind (which might be acceptable).
Not clear if the profile needs to be fixed or the firmware.